pyproject.toml
[tool.poetry]
name = "ratdream"
version = "0.1.0"
description = "Бекенд для цифрового прорыва"
authors = ["Timur <100406268+Timik232@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
protobuf = "<=3.17.3"
grpcio = "<=1.46.0"
grpcio-tools = "<=1.30.0"
pika = "<=1.2.0"
pyedflib = "<=0.1.38"
pandas = "<=1.3.5"
joblib = "<=1.2.0"
scikit-learn = "==1.2.2"
tqdm = "<=4.62.3"
statsmodels = "<=0.12.2"
scipy = "<=1.7.3"

[tool.poetry.group.dev.dependencies]
pyinstaller = "<=4.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
